break out test harness into its own file
