+++ test/dependency-manager/spec/component-spec.unit.test.ts
@@ -1,74 +1,14 @@
-import deepmerge from 'deepmerge';
-import fs from 'fs-extra';
 import mock_fs from 'mock-fs';
 import moxios from 'moxios';
-import path from 'path';
 import sinon from 'sinon';
 import Register from '../../../src/commands/register';
 import PortUtil from '../../../src/common/utils/port';
-import { buildConfigFromYml, dumpToYml, parseSourceYml, Slugs } from '../../../src/dependency-manager/src';
-
-const lsDirectories = (dir: string) => {
-  return fs.readdirSync(path.resolve(dir), { withFileTypes: true })
-    .filter(dirent => dirent.isDirectory())
-    .map(dirent => dirent.name);
-}
-
-export interface RecursivePartialsTree {
-  children: { [key: string]: RecursivePartialsTree };
-  partials: Record<string, any>[];
-}
-
-const recursiveBuildTree = (dir: string): RecursivePartialsTree => {
-  const files = readFiles(dir);
-  const parsed_source_ymls = files.map(s => parseSourceYml(s) as object);
-  const directories = lsDirectories(dir);
-  const children = directories.reduce((obj: { [key: string]: RecursivePartialsTree }, dir_name: string) => {
-    obj[dir_name] = recursiveBuildTree(`${dir}/${dir_name}`);
-    return obj;
-  }, {});
-  return {
-    children,
-    partials: parsed_source_ymls
-  };
-}
-
-const recursiveMergeTree = (tree: RecursivePartialsTree): object[] => {
-  if (!Object.keys(tree.children).length) {
-    return tree.partials;
-  }
-
-  return Object.entries(tree.children)
-    .map(([child_key, child_tree]) => {
-      const child_partials = recursiveMergeTree(child_tree);
-
-      const merged_partials = tree.partials.map(current_partial => {
-        return child_partials.map(child_partial => {
-          if (child_key.startsWith('_')) {
-            let accum = current_partial;
-            for (const current_key of Object.keys(accum)) {
-              accum[current_key] = deepmerge(accum[current_key], { [child_key.replace('_', '')]: child_partial });
-            }
-            return accum;
-          } else {
-            return deepmerge(current_partial, { [child_key]: child_partial }) as object
-          }
-        });
-      }).reduce((accumulator, value) => accumulator.concat(value), []);
-
-      return merged_partials;
-    }).reduce((accumulator, value) => accumulator.concat(value), []);
-}
-
-const readFiles = (directory: string): string[] => {
-  return fs.readdirSync(directory)
-    .filter(f => fs.lstatSync(path.resolve(directory, f)).isFile())
-    .map(f => {
-      return fs.readFileSync(path.resolve(directory, f), 'utf-8');
-    });
-};
+import { buildConfigFromYml, dumpToYml, Slugs } from '../../../src/dependency-manager/src';
+import { loadAllTestSpecCombinations } from './partials/spec-test-harness';
 
 describe('component spec unit test', function () {
+  const all_spec_combinations = loadAllTestSpecCombinations();
+
   beforeEach(async () => {
 
     // Stub the logger
@@ -95,16 +35,11 @@ describe('component spec unit test', function () {
     moxios.uninstall();
   });
 
-  const root_partials_dir = `test/dependency-manager/spec/partials/root`;
-
   describe(`recursively test partial architect components`, async () => {
-    const tree = recursiveBuildTree(root_partials_dir);
-
-    const all_components = recursiveMergeTree(tree);
 
     it(`loadSourceYmlFromPathOrReject loads valid file`, async () => {
       const errors = [];
-      for (const component of all_components) {
+      for (const component of all_spec_combinations) {
         const source_yml = dumpToYml(component);
         buildConfigFromYml(source_yml, Slugs.DEFAULT_TAG);
       }
